+++ src/Dependencies/PooledObjects/ArrayBuilder.cs
@@ -347,23 +347,11 @@ namespace Microsoft.CodeAnalysis.PooledObjects
             return builder;
         }
 
-        public static ArrayBuilderDisposer GetInstance(out ArrayBuilder<T> instance)
-        {
-            instance = GetInstance();
-            return new ArrayBuilderDisposer(instance);
-        }
-
         public static ArrayBuilder<T> GetInstance(int capacity)
         {
             var builder = GetInstance();
             builder.EnsureCapacity(capacity);
             return builder;
-        }
-
-        public static ArrayBuilderDisposer GetInstance(int capacity, out ArrayBuilder<T> instance)
-        {
-            instance = GetInstance(capacity);
-            return new ArrayBuilderDisposer(instance);
         }
 
         public static ArrayBuilder<T> GetInstance(int capacity, T fillWithValue)
@@ -377,12 +365,6 @@ namespace Microsoft.CodeAnalysis.PooledObjects
             }
 
             return builder;
-        }
-
-        public static ArrayBuilderDisposer GetInstance(int capacity, T fillWithValue, out ArrayBuilder<T> instance)
-        {
-            instance = GetInstance(capacity, fillWithValue);
-            return new ArrayBuilderDisposer(instance);
         }
 
         public static ObjectPool<ArrayBuilder<T>> CreatePool()
@@ -581,27 +563,5 @@ namespace Microsoft.CodeAnalysis.PooledObjects
             set.Free();
             return result.ToImmutableAndFree();
         }
-
-        #region IDisposable Support
-        internal struct ArrayBuilderDisposer : IDisposable
-        {
-            private ArrayBuilder<T> _pooledObject;
-
-            public ArrayBuilderDisposer(ArrayBuilder<T> instance)
-            {
-                _pooledObject = instance;
-            }
-
-            public void Dispose()
-            {
-                var pooledObject = _pooledObject;
-                if (pooledObject != null)
-                {
-                    pooledObject.Free();
-                    _pooledObject = null;
-                }
-            }
-        }
-        #endregion
     }
 }
